requirements.txt
# Core Pydantic AI Framework
pydantic-ai>=0.0.13
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment Configuration
python-dotenv>=1.0.0

# HTTP Clients
httpx>=0.27.0
requests>=2.31.0

# Image Processing
Pillow>=10.0.0

# Retry Logic
tenacity>=8.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-mock>=3.12.0

# Development Dependencies
black>=23.0.0
ruff>=0.1.0
mypy>=1.7.0
